requirements.txt
appnope==0.1.2
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
bottleneck==1.3.4
certifi==2022.5.18.1
cffi==1.15.0
click==8.0.4
dawg-python==0.7.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
dostoevsky==0.6.0
entrypoints==0.4
executing==0.8.3
fastapi[all]==0.78.0
fasttext==0.9.2
intel-openmp==2021.4.0
intervaltree==3.1.0
ipykernel==6.9.1
ipymarkup==0.9.0
ipython==8.3.0
ipython_genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter_client==7.2.2
jupyter_console==6.4.3
jupyter_core==4.10.0
jupyterlab_pygments==0.1.2
jupyterlab_widgets==1.0.0
markupsafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl==2022.1.0
natasha==1.4.0
navec==0.10.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
numexpr==2.8.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.1.2
plotly==5.8.0
prometheus_client==0.13.1
prompt_toolkit==3.0.20
ptyprocess==0.7.0
pure_eval==0.2.2
pybind11==2.9.2
pycparser==2.21
pydantic==1.9.1
pygments==2.11.2
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy2==0.9.1
pyparsing==3.0.4
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.3.0
qtpy==2.0.1
razdel==0.5.0
regex==2022.3.15
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
send2trash==1.8.0
setuptools==61.2.0
sip==6.6.1
six==1.16.0
slovnet==0.5.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlalchemy==1.4.37
stack_data==0.2.0
tenacity==8.0.1
terminado==0.13.1
translate==3.6.1
testpath==0.6.0
threadpoolctl==3.1.0
tk==0.1.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
translate==3.6.1
typing-extensions==4.1.1
tzdata==2022.1
uvicorn==0.17.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.5.2
yargy==0.15.0
joblib==1.1.0
htmldate==1.2.1
openpyxl==3.0.10
newspaper3k==0.2.8
pyOpenSSL
tld
whois